Perception: [fixed]:DAG bugs, Visualizer bugs, Sensor Content bugs. [enable]: Motion Service with Visualization (#3569)
* Perception: comments on motion buffer generation

* Perception: Revert previous code refactor, causing mistakes on subnode_map_. Subnodes should be loaded in-order.

* Perception: fix overfloating in timestamp string generator inside CommonSharedData

* Perception: moition manager changes to provide buffer for motions ahead of images

* Perception: moition manager visualization enable and correction

* Perception: fix previous bugs in code reversion, In visualization frame content, key and timestamp should be consistent.

+++ modules/perception/onboard/common_shared_data.h
@@ -51,8 +51,8 @@ struct CommonSharedDataKey {
       : timestamp(ts), device_id(id) {}
   virtual std::string ToString() const {
     return device_id +
-           (boost::format("%d") %
-            static_cast<int>(timestamp * FLAGS_stamp_enlarge_factor))
+           (boost::format("%ld") %
+            static_cast<int64_t>(timestamp * FLAGS_stamp_enlarge_factor))
                .str();
   }
   double timestamp = 0.0;
@@ -203,6 +203,7 @@ bool CommonSharedData<M>::Add(const CommonSharedDataKey &key,
                               const SharedDataPtr<M> &data) {
   // update latest_timestamp for SharedData
   latest_timestamp_ = std::max(latest_timestamp_, key.timestamp);
+  // latest_timestamp_ = key.timestamp;
   return Add(key.ToString(), data);
 }
